chore: Refactor brand store state management
