+++ static/js/brand-store/utils/index.ts
@@ -1,22 +1,16 @@
 import checkModelNameExists from "./checkModelNameExists";
-import getFilteredModels from "./getFilteredModels";
-import getFilteredPolicies from "./getFilteredPolicies";
 import isClosedPanel from "./isClosedPanel";
 import maskString from "./maskString";
 import setPageTitle from "./setPageTitle";
-import getFilteredSigningKeys from "./getFilteredSigningKeys";
 import checkSigningKeyExists from "./checkSigningKeyExists";
 import sortByDateDescending from "./sortByDateDescending";
 import getPolicies from "./getPolicies";
 
 export {
   checkModelNameExists,
-  getFilteredModels,
-  getFilteredPolicies,
   isClosedPanel,
   maskString,
   setPageTitle,
-  getFilteredSigningKeys,
   checkSigningKeyExists,
   sortByDateDescending,
   getPolicies,
